feat(udf): asc::compile_assemblyscript library API behind asc-udf feature
Promote the proven wasmtime-backed asc/binaryen shim from an example into a
reusable library function: compile_assemblyscript(source) -> Result<Vec<u8>>.

A persistent compiler thread owns the QuickJS runtime for the process lifetime:
it parses the 15 MB bundle and runs binaryen init once, then serves every
compile. This amortises init and sidesteps the JS<->Rust GC teardown assertion
(the runtime is never freed) without leaking per compile. Bundle is located via
FERROSA_ASC_BUNDLE; rquickjs/tokio are gated behind the asc-udf feature so
default builds are unaffected. Live-infra test compiles add and runs it (=>5).
